+++ eNode.c
@@ -1,7 +1,7 @@
 /****************************************************************************
  *                      Embedded node protocol(ENP)                         *
  ****************************************************************************/
-#include "enpapi.h"
+#include "enpHelper.h"
 #include "stddef.h"
 
 // nodes list
