+++ enCANProt.h
@@ -3,7 +3,7 @@
  ****************************************************************************/
 #ifndef SMP_H_
 #define SMP_H_
-#include "enpapi.h"
+#include "enpHelper.h"
 #include "stdint.h"
 
 extern uint32_t SMP_Error;
