+++ enpHelper.c
@@ -1,7 +1,7 @@
 /****************************************************************************
  *                      Embedded node protocol(ENP)                         *
  ****************************************************************************/
-#include "enpapi.h"
+#include "enpHelper.h"
 
 // Таблица стандартного CRC-16 (ARC)
 static const uint16_t CRC16_Table[256] = {
